+++ presidio-image-redactor/presidio_image_redactor/__init__.py
@@ -5,10 +5,12 @@ from .ocr import OCR
 from .tesseract_ocr import TesseractOCR
 from .bbox import BboxProcessor
 from .image_analyzer_engine import ImageAnalyzerEngine
+from .qr_image_analyzer_engine import QRImageAnalyzerEngine
 from .image_redactor_engine import ImageRedactorEngine
 from .image_pii_verify_engine import ImagePiiVerifyEngine
 from .dicom_image_redactor_engine import DicomImageRedactorEngine
 from .dicom_image_pii_verify_engine import DicomImagePiiVerifyEngine
+
 
 # Set up default logging (with NullHandler)
 logging.getLogger("presidio-image-redactor").addHandler(logging.NullHandler())
@@ -18,6 +20,7 @@ __all__ = [
     "TesseractOCR",
     "BboxProcessor",
     "ImageAnalyzerEngine",
+    "QRImageAnalyzerEngine",
     "ImageRedactorEngine",
     "ImagePiiVerifyEngine",
     "DicomImageRedactorEngine",